[Silver III] Title: 주유소, Time: 372 ms, Memory: 35644 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -16,12 +16,12 @@ public class Main {
             oilPrice[i]=Integer.parseInt(st.nextToken());
         }
         int cheapPriceCity=Integer.MAX_VALUE;
-        int answer = 0;
+        long answer = 0;
         for(int i=0;i<edge.length;i++){
             if(oilPrice[i]<cheapPriceCity){
                 cheapPriceCity=oilPrice[i];
             }
-            answer+=edge[i]*cheapPriceCity;
+            answer+=(long)edge[i]*(long)cheapPriceCity;
         }
         System.out.println(answer);
     }
